agentcore_runtime/requirements.txt
strands-agents
bedrock-agentcore
boto3